requirements.txt
requests==2.24.0
pyqt5==5.15.0
black==20.8b1
pyqt5-tools==5.15.0.1.7.1
CherryPy==18.6.0
python-dotenv==0.14.0
